Move "recent" types to file-system/models
