+++ src/stores/recent-files-store.ts
@@ -1,7 +1,7 @@
 import { create } from "zustand";
 import { persist } from "zustand/middleware";
 import { immer } from "zustand/middleware/immer";
-import type { RecentFile, RecentFilesStore } from "../types/recent-files";
+import type { RecentFile, RecentFilesStore } from "../file-system/models/recent-files";
 import { calculateFrecencyScore } from "../utils/frecency";
 
 const MAX_RECENT_FILES = 50;
