+++ src/stores/recent-folders-store.ts
@@ -1,7 +1,7 @@
 import { create } from "zustand";
 import { persist } from "zustand/middleware";
 import { immer } from "zustand/middleware/immer";
-import type { RecentFolder } from "../types/recent-folders";
+import type { RecentFolder } from "../file-system/models/recent-folders";
 
 interface RecentFoldersState {
   recentFolders: RecentFolder[];
